+++ src/ast/mod.rs
@@ -1,24 +1,23 @@
 #[macro_use]
 mod parser_util;
 mod parsers;
+pub mod punctuated;
 
 use crate::tokenizer::{Symbol, Token, TokenKind, TokenReference, TokenType};
 use full_moon_derive::{Node, Visit};
 use generational_arena::Arena;
 use itertools::Itertools;
-use parser_util::{
-    InternalAstError,
-    OneOrMore,
-    Parser,
-    ParserState,
-    ZeroOrMore,
-    ZeroOrMoreDelimited,
-};
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 use std::fmt;
 use std::iter::FromIterator;
 use std::sync::Arc;
+
+use parser_util::{
+    InternalAstError, OneOrMore, Parser, ParserState, ZeroOrMore, ZeroOrMoreDelimited,
+};
+
+use punctuated::{Pair, Punctuated};
 
 /// A block of statements, such as in if/do/etc block
 #[derive(Clone, Debug, PartialEq, Node, Visit)]
@@ -54,7 +53,7 @@ pub enum LastStmt<'a> {
         #[cfg_attr(feature = "serde", serde(borrow))]
         token: TokenReference<'a>,
         /// The values being returned
-        returns: Vec<Expression<'a>>,
+        returns: Punctuated<'a, Expression<'a>>,
     },
 }
 
@@ -256,7 +255,7 @@ pub enum Index<'a> {
 pub enum FunctionArgs<'a> {
     #[cfg_attr(feature = "serde", serde(borrow))]
     /// Used when a function is called in the form of `call(1, 2, 3)`
-    Parentheses(Vec<Expression<'a>>),
+    Parentheses(Punctuated<'a, Expression<'a>>),
     /// Used when a function is called in the form of `call "foobar"`
     String(TokenReference<'a>),
     /// Used when a function is called in the form of `call { 1, 2, 3 }`
@@ -310,23 +309,23 @@ impl<'a> NumericFor<'a> {
 pub struct GenericFor<'a> {
     #[cfg_attr(feature = "serde", serde(borrow))]
     for_token: TokenReference<'a>,
-    names: Vec<TokenReference<'a>>,
-    expr_list: Vec<Expression<'a>>,
+    names: Punctuated<'a, TokenReference<'a>>,
+    expr_list: Punctuated<'a, Expression<'a>>,
     block: Block<'a>,
     end_token: TokenReference<'a>,
 }
 
 impl<'a> GenericFor<'a> {
-    /// An iterator over the names used in a for loop
+    /// Returns the [`Punctuated`](punctuated/struct.Punctuated.html) sequence of names
     /// In `for index, value in pairs(list) do`, iterates over `index` and `value`
-    pub fn iter_names(&self) -> impl Iterator<Item = &TokenReference<'a>> {
-        self.names.iter()
+    pub fn names(&self) -> &Punctuated<'a, TokenReference<'a>> {
+        &self.names
     }
 
-    /// An iterator over the expression used in a for loop
+    /// Returns the [`Punctuated`](punctuated/struct.Punctuated.html) sequence of the expressions looped over
     /// In `for index, value in pairs(list) do`, iterates over `pairs(list)`
-    pub fn iter_expr_list(&self) -> impl Iterator<Item = &Expression<'a>> {
-        self.expr_list.iter()
+    pub fn expr_list(&self) -> &Punctuated<'a, Expression<'a>> {
+        &self.expr_list
     }
 
     /// The code inside the for loop
@@ -487,7 +486,7 @@ pub enum Parameter<'a> {
     /// The `...` vararg syntax, such as `function x(...)`
     Ellipse(TokenReference<'a>),
     /// A name parameter, such as `function x(a, b, c)`
-    Name(TokenReference<'a>),
+    Name(Pair<'a, TokenReference<'a>>),
 }
 
 /// A suffix in certain cases, such as `:y()` in `x:y()`
@@ -539,19 +538,21 @@ pub enum Var<'a> {
 #[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 pub struct Assignment<'a> {
     #[cfg_attr(feature = "serde", serde(borrow))]
-    var_list: Vec<Var<'a>>,
-    expr_list: Vec<Expression<'a>>,
+    var_list: Punctuated<'a, Var<'a>>,
+    expr_list: Punctuated<'a, Expression<'a>>,
 }
 
 impl<'a> Assignment<'a> {
-    /// An iterator over the expressions being assigned, the `1, 2` part of `x, y["a"] = 1, 2`
-    pub fn iter_expr_list(&self) -> impl Iterator<Item = &Expression<'a>> {
-        self.expr_list.iter()
+    /// Returns the [`Punctuated`](punctuated/struct.Punctuated.html) sequence over the expressions being assigned.
+    /// This is the the `1, 2` part of `x, y["a"] = 1, 2`
+    pub fn expr_list(&self) -> &Punctuated<'a, Expression<'a>> {
+        &self.expr_list
     }
 
-    /// An iterator over the variables being assigned to, the `x, y["a"]` part of `x, y["a"] = 1, 2`
-    pub fn iter_var_list(&self) -> impl Iterator<Item = &Var<'a>> {
-        self.var_list.iter()
+    /// Returns the [`Punctuated`](punctuated/struct.Punctuated.html) sequence over the variables being assigned to.
+    /// This is the `x, y["a"]` part of `x, y["a"] = 1, 2`
+    pub fn var_list(&self) -> &Punctuated<'a, Var<'a>> {
+        &self.var_list
     }
 }
 
@@ -583,24 +584,29 @@ impl<'a> LocalFunction<'a> {
 pub struct LocalAssignment<'a> {
     #[cfg_attr(feature = "serde", serde(borrow))]
     local_token: TokenReference<'a>,
-    name_list: Vec<TokenReference<'a>>,
-    expr_list: Vec<Expression<'a>>,
+    name_list: Punctuated<'a, TokenReference<'a>>,
+    expr_list: Punctuated<'a, Expression<'a>>,
 }
 
 impl<'a> LocalAssignment<'a> {
-    /// An iterator over the expressions being assigned, the `1, 2` part of `local x, y = 1, 2`
-    pub fn iter_expr_list(&self) -> impl Iterator<Item = &Expression<'a>> {
-        self.expr_list.iter()
+    /// Returns the [`Punctuated`](punctuated/struct.Punctuated.html) sequence of the expressions being assigned.
+    /// This is the `1, 2` part of `local x, y = 1, 2`
+    pub fn expr_list(&self) -> &Punctuated<'a, Expression<'a>> {
+        &self.expr_list
     }
 
-    /// An iterator over the names being assigned to, the `x, y` part of `local x, y = 1, 2`
-    pub fn iter_name_list(&self) -> impl Iterator<Item = &TokenReference<'a>> {
-        self.name_list.iter()
+    /// Returns the [`Punctuated`](punctuated/struct.Punctuated.html) sequence of names being assigned to.
+    /// This is the `x, y` part of `local x, y = 1, 2`
+    pub fn name_list(&self) -> &Punctuated<'a, TokenReference<'a>> {
+        &self.name_list
     }
 
-    /// A mutable iterator over the names being assigned to, the `x, y` part of `local x, y = 1, 2`
-    pub fn iter_name_list_mut(&mut self) -> impl Iterator<Item = &mut TokenReference<'a>> {
-        self.name_list.iter_mut()
+    /// Returns a mutable [`Punctuated`](punctuated/struct.Punctuated.html) sequence of names being assigned to.
+    /// This is the `x, y` part of `local x, y = 1, 2`
+    pub fn name_list_mut(
+        &mut self,
+    ) -> &mut Punctuated<'a, TokenReference<'a>> {
+        &mut self.name_list
     }
 }
 
@@ -641,7 +647,7 @@ impl<'a> FunctionCall<'a> {
 #[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 pub struct FunctionName<'a> {
     #[cfg_attr(feature = "serde", serde(borrow))]
-    names: Vec<TokenReference<'a>>,
+    names: Punctuated<'a, TokenReference<'a>>,
     colon_name: Option<TokenReference<'a>>,
 }
 
@@ -651,9 +657,10 @@ impl<'a> FunctionName<'a> {
         self.colon_name.as_ref()
     }
 
-    /// An iterator over the names used when defining the function, the `x.y.z` part of `function x.y.z() end`
-    pub fn iter_names(&self) -> impl Iterator<Item = &TokenReference<'a>> {
-        self.names.iter()
+    /// Returns the [`Punctuated`](punctuated/struct.Punctuated.html) sequence over the names used when defining the function.
+    /// This is the `x.y.z` part of `function x.y.z() end`
+    pub fn names(&self) -> &Punctuated<'a, TokenReference<'a>> {
+        &self.names
     }
 }